Rewrite
Rewrote most of the code as I began to figure out how I wanted to use vexflow
and structure the code. It is still in a partial state as of this commit.
However, it does renderer and randomly re-order the measures. The next major
addition is to sort out the layout so that it is possible to simply pass in the
measures as sets of notes and automatically figure out how to lay the measures
out based on the available width of the SVG that is being rendered to.