index: changed svg to div
Using the svg element was the wrong thing to do and cause a clipping issue. The
'canvas' element will be the parent of an SVG element created by Raphael.
Embedding an SVG element inside another SVG element causes some confusion when
it comes to determine the size of the element.